Rename 'components/MonitoringConfirmedCasesChart.vue' to 'components/index/CardsMonitoring/MonitoringConfirmedCasesNumber/Chart.vue'
