+++ components/index/CardsMonitoring/MonitoringConfirmedCasesNumber/Chart.vue
